# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /results/summary:
    get:
      operationId: getResultsSummary
      tags: [results]
      summary: Get overall and per-sport model performance summary
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Results summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultsSummaryResponse"
  /results/roi:
    get:
      operationId: getResultsRoi
      tags: [results]
      summary: Get win%, units, and ROI broken down by sport and rating tier
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: ROI breakdown by rating and sport
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultsRoiResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    masterApiKey:
      type: apiKey
      in: header
      name: X-Master-Key
      description: Internal master key. An authenticated admin session token is also accepted by the server.
  schemas:
...
    OverallStat:
      type: object
      properties:
        wins:
          type: integer
        losses:
          type: integer
        pushes:
          type: integer
        totalPicks:
          type: integer
        winRate:
          type: number
        unitsWonLost:
          type: number
      required: [wins, losses, pushes, totalPicks, winRate, unitsWonLost]
    SportStat:
      type: object
      properties:
        sport:
          type: string
        wins:
          type: integer
        losses:
          type: integer
        pushes:
          type: integer
        totalPicks:
          type: integer
        winRate:
          type: number
        unitsWonLost:
          type: number
        currentStreak:
          type: integer
        currentStreakDir:
          type: string
      required: [sport, wins, losses, pushes, totalPicks, winRate, unitsWonLost, currentStreak, currentStreakDir]
    RecentResult:
      type: object
      properties:
        pickId:
          type: integer
        sport:
          type: string
        awayTeamAbbr:
          type: string
        homeTeamAbbr:
          type: string
        awayScore:
          type: integer
          nullable: true
        homeScore:
          type: integer
          nullable: true
        pick:
          type: string
        odds:
          type: integer
          nullable: true
        unitsRisked:
          type: number
          nullable: true
        unitsWonLost:
          type: number
        result:
          type: string
        gameDate:
          type: string
        gradedAt:
          type: string
          nullable: true
        modelId:
          type: string
        modelVersionId:
          type: integer
        provenance:
          type: string
          enum: [preCutoverOfficial, v4Official]
      required: [pickId, sport, awayTeamAbbr, homeTeamAbbr, pick, unitsWonLost, result, gameDate, modelId, modelVersionId, provenance]
    RecordSegment:
      type: object
      description: Official wager results grouped by immutable persisted publication and model provenance.
      properties:
        wins:
          type: integer
        losses:
          type: integer
        pushes:
          type: integer
        totalPicks:
          type: integer
        winRate:
          type: number
        unitsWonLost:
          type: number
        unitsRisked:
          type: number
        roi:
          type: number
      required: [wins, losses, pushes, totalPicks, winRate, unitsWonLost, unitsRisked, roi]
    ResultsSummaryResponse:
      type: object
      properties:
        period:
          type: string
        overall:
          $ref: "#/components/schemas/OverallStat"
        recordSegments:
          type: object
          properties:
            preCutoverOfficial:
              $ref: "#/components/schemas/RecordSegment"
            v4Official:
              $ref: "#/components/schemas/RecordSegment"
          required: [preCutoverOfficial, v4Official]
        bySport:
          type: array
          items:
            $ref: "#/components/schemas/SportStat"
        recentResults:
          type: array
          items:
            $ref: "#/components/schemas/RecentResult"
        dataAsOf:
          type: string
      required: [overall, recordSegments, bySport]
    RoiEntry:
      type: object
      properties:
        key:
          type: string
        wins:
          type: integer
        losses:
          type: integer
        pushes:
          type: integer
        totalPicks:
          type: integer
        winRate:
          type: number
        unitsWonLost:
          type: number
        unitsRisked:
          type: number
        roi:
          type: number
      required: [key, wins, losses, pushes, totalPicks, winRate, unitsWonLost, unitsRisked, roi]
    RoiByRatingEntry:
      allOf:
        - $ref: "#/components/schemas/RoiEntry"
        - type: object
          properties:
            recommendation:
              type: string
          required: [recommendation]
    RoiBySportEntry:
      allOf:
        - $ref: "#/components/schemas/RoiEntry"
        - type: object
          properties:
            sport:
              type: string
          required: [sport]
    RoiBySportAndRatingEntry:
      allOf:
        - $ref: "#/components/schemas/RoiEntry"
        - type: object
          properties:
            sport:
              type: string
            recommendation:
              type: string
          required: [sport, recommendation]
    ResultsRoiResponse:
      type: object
      properties:
        period:
          type: string
        byRating:
          type: array
          items:
            $ref: "#/components/schemas/RoiByRatingEntry"
        bySport:
          type: array
          items:
            $ref: "#/components/schemas/RoiBySportEntry"
        bySportAndRating:
          type: array
          items:
            $ref: "#/components/schemas/RoiBySportAndRatingEntry"
        dataAsOf:
          type: string
      required: [byRating, bySport, bySportAndRating]
...