# lib/api-zod/src/generated/types/resultsSummaryResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OverallStat } from './overallStat';
import type { RecentResult } from './recentResult';
import type { ResultsSummaryResponseRecordSegments } from './resultsSummaryResponseRecordSegments';
import type { SportStat } from './sportStat';

export interface ResultsSummaryResponse {
  period?: string;
  overall: OverallStat;
  recordSegments: ResultsSummaryResponseRecordSegments;
  bySport: SportStat[];
  recentResults?: RecentResult[];
  dataAsOf?: string;
}